Change representation for parsers
Representing parsers as functions seemed a good idea at the time I
implemented it, but it turned out to be a bad idea, since it caused a
problem:
Because the construction of a function-value requires the enclosing
context (dictionary), auxiliary words that are defined later are
unknown in such a function value. In combination with the evaluation
of list literals in words beeing defined this caused problems.